Added an abstract layout
